fix: umount issue leaves global lock (#95)
We saw an issue where one Unmount operation suspended, and because it
had the `volumeLock` locked, subsequent operations also all suspended
because they could not obtain the lock.

We have refactored locking access to the `volumes` map so that fewer
operations happen under lock, and a deadlock is now less likely.

Co-authored-by: Konstantin Kiess <kkiess@vmware.com>