chore: refactor Mount() for clarity
It looks like the Mount() method once used goroutines, and while this
was removed, the complexity was not. This change does not change any
behavior, but removes overly-complex structure.

At the same time, VolumeDriver.volumes was changes to a map of value
types (rather than pointers) to make it explict when updates were
happening, as anyone with access to the pointer could make changes
without the protection of a lock, but when a value, it's necessary to do
an explicit Put() to update the map, so access to data is more
controlled.

There are still some potential data races in this code, for instance if
two Mount() calls happen at the same time, then the mount counter might
be incremented only once. The risk from this is low, but they should be
fixed in the future.